implementaed signal monitor
