+++ client/src/androidMain/kotlin/org/apptank/horus/client/HorusConfigurator.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import org.apptank.horus.client.di.HorusContainer
 import com.russhwolf.settings.SharedPreferencesSettings
 import org.apptank.horus.client.config.HorusConfig
-import java.io.File
+import org.apptank.horus.client.connectivity.NetworkValidator
 
 /**
  * HorusConfigurator is a class that allows you to configure the Horus SDK.
