+++ gradle.properties
@@ -6,7 +6,7 @@ org.gradle.configuration-cache=true
 # App version
 lib.groupId = org.apptank.horus
 lib.artifactId = client
-lib.version = 0.14.0-alpha12
+lib.version = 0.14.0-alpha19
 
 #Kotlin
 kotlin.code.style=official
